Support gz-compressed input files.
As an unavoidable side effect, this change turns off universal newline
mode on inputs. Since all input flows through the csv module, and the
use of universal newline input was not consistent across csvkit in any
case, my understanding is that turning it off should not cause a
regression. Someone more familiar with the subtleties of newline
handling in CSV may want to chime in.