remove $ for more beginner friendly docs
there was a NICAR thread where someone new to python was trying to use this tutorial, and copy pasted an entire line, including the '$'. google doesn't recognize '$' so search results for '$: command not found' aren't helpful. this is a common issue for python beginners.

makes me sad to think of all the new folks who might be missing out on the wonders of csvkit, b/c this getting started tutorial isn't copy/paste friendly! here's an easy fix.